+++ errors.go
@@ -18,9 +18,10 @@ var (
 	errAuthnWebhookResponse = errors.New("AuthnWebhookResponseError")
 	errAuthnWebhookReject   = errors.New("AuthnWebhookReject")
 
+	errDisconnectWebhook         = errors.New("DisconnectWebhookError")
 	errDisconnectWebhookResponse = errors.New("DisconnectWebhookResponseError")
 
 	errRoomFull = errors.New("RoomFull")
 	// 想定外のエラー
-	errInternalServer = errors.New("InternalServerError")
+	// errInternalServer = errors.New("InternalServerError")
 )