+++ CHANGES.md
@@ -14,24 +14,45 @@
 - [ADD] register メッセージで key と signalingKey のどちらかを指定できるようにする
     - signalingKey が優先される
     - 将来的に signalingKey のみになる
+    - @voluntas
 - [ADD] accept メッセージで isExistUser 以外に isExistClient を送るようにする
     - 将来的に isExistClient のみになる
+    - @voluntas
 - [ADD] 切断時にウェブフック通知を飛ばせるようにする
     - disconnect_webhook_url を設定
+    - @voluntas @Hexa
 - [ADD] signaling.log を追加する
+    - @voluntas @Hexa
 - [ADD] webhook.log を追加する
+    - @voluntas @Hexa
 - [ADD] register メッセージで ayameClient / environment / libwebrtc の情報を追加する
     - 認証ウェブフック通知で含まれるようにする
+    - @voluntas
+- [ADD] type: accept 時に connectionId を払い出すようにする
+    - @voluntas
 - [CHANGE] コードベースを変更する
+    - @voluntas @Hexa
 - [CHANGE] addr を listen_ipv4_address に変更する
+    - @voluntas
 - [CHANGE] port を listen_port_number に変更する
+    - @voluntas
 - [CHANGE] allow_origin 設定を削除する
+    - @voluntas
 - [CHANGE] ロガーを zerolog に変更する
+    - @voluntas
 - [CHANGE] ログローテーションを lumberjack に変更する
-- [CHANGE] 同一クライアント ID での同一ルームへの接続をできなくする
+    - @voluntas
 - [CHANGE] サンプルを削除する
+    - @voluntas
 - [CHANGE] 登録済みのあとに WebSocket 切断した場合、 type: bye を送信するようにする
+    - @voluntas @Hexa
 - [CHANGE] ウェブフックの戻り値のステータスコード 200 以外はエラーにする
+    - @voluntas @Hexa
 - [CHANGE] ウェブフックの JSON のキーを snake_case から camelCase にする
+    - @voluntas
+- [CHANGE] clientId をオプション化する
+    - @voluntas
 - [FIX] サーバ側の切断の WS の終了処理を適切に行う
+    - @voluntas @Hexa
 - [FIX] ウェブソケットの最大メッセージを 1MB に制限する
+    - @voluntas